refactor(admin): split handler.go by domain + extract route registration
internal/admin/handler.go was 2400 lines hosting endpoints for ten
distinct domains (usage, audit, models, providers, budgets, aliases,
auth keys, model overrides, guardrails, workflows). Split methods into
peer files in the same package, mirroring the existing per-domain test
file layout.

  handler.go (513)               struct, options, NewHandler, runtime
                                 refresh types, parse/error helpers
  errors.go (141)                feature/validation/write error helpers,
                                 deactivateByID, deleteByName, decode helpers
  routes.go (68)                 RegisterRoutes(RouteRegistrar) — admin
                                 RouteRegistrar interface so callers can
                                 mount under any route group
  handler_usage.go (398)         UsageSummary, DailyUsage, UsageByModel,
                                 UsageByUserPath, UsageLog, CacheOverview,
                                 RecalculateUsagePricing + recalc helpers/types
  handler_audit.go (205)         AuditLog, AuditConversation, auditLogResponse
  handler_models.go (129)        ListModels, ListCategories, DashboardConfig
  handler_providers.go (169)     ProviderStatus, RefreshRuntime,
                                 buildProviderStatusResponse, classifyProviderStatus
  handler_budgets.go (385)       budget endpoints + types + helpers
  handler_aliases.go (91)        alias endpoints + types
  handler_authkeys.go (81)       auth-key endpoints + types
  handler_model_overrides.go (79) model-override endpoints + types
  handler_guardrails.go (118)    guardrail endpoints + types
  handler_workflows.go (246)     workflow endpoints + types + validation helpers

internal/server/http.go: replaced the 40-line inline admin route block
with `cfg.AdminHandler.RegisterRoutes(e.Group("/admin/api/v1"))`. Route
registration now lives next to the methods it wires up.

Pure relocation; full test suite passes.
