+++ internal/admin/handler.go
@@ -3,14 +3,9 @@ package admin
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
-	"fmt"
 	"log/slog"
 	"net/http"
-	"net/url"
-	"slices"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
@@ -507,1894 +502,12 @@ func requestIDFromAdminContextOrHeader(req *http.Request) string {
 	return strings.TrimSpace(req.Header.Get("X-Request-ID"))
 }
 
-// UsageSummary handles GET /admin/api/v1/usage/summary
-//
-// @Summary      Get usage summary
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Param        days        query     int     false  "Number of days (default 30)"
-// @Param        start_date  query     string  false  "Start date (YYYY-MM-DD)"
-// @Param        end_date    query     string  false  "End date (YYYY-MM-DD)"
-// @Param        user_path   query     string  false  "Filter by tracked user path subtree"
-// @Param        cache_mode  query     string  false  "Cache mode filter: uncached, cached, all (default uncached)"
-// @Success      200  {object}  usage.UsageSummary
-// @Failure      400  {object}  core.GatewayError
-// @Failure      401  {object}  core.GatewayError
-// @Router       /admin/api/v1/usage/summary [get]
-func (h *Handler) UsageSummary(c *echo.Context) error {
-	if h.usageReader == nil {
-		return c.JSON(http.StatusOK, usage.UsageSummary{})
-	}
-
-	params, err := parseUsageParams(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	summary, err := h.usageReader.GetSummary(c.Request().Context(), params)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	return c.JSON(http.StatusOK, summary)
-}
-
-func usageSliceResponse[T any](
-	c *echo.Context,
-	reader usage.UsageReader,
-	fetch func(context.Context, usage.UsageQueryParams) ([]T, error),
-) error {
-	if reader == nil {
-		return c.JSON(http.StatusOK, []T{})
-	}
-
-	params, err := parseUsageParams(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	values, err := fetch(c.Request().Context(), params)
-	if err != nil {
-		return handleError(c, err)
-	}
-	if values == nil {
-		values = []T{}
-	}
-	return c.JSON(http.StatusOK, values)
-}
-
-// DailyUsage handles GET /admin/api/v1/usage/daily
-//
-// @Summary      Get usage breakdown by period
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Param        days        query     int     false  "Number of days (default 30)"
-// @Param        start_date  query     string  false  "Start date (YYYY-MM-DD)"
-// @Param        end_date    query     string  false  "End date (YYYY-MM-DD)"
-// @Param        interval    query     string  false  "Grouping interval: daily, weekly, monthly, yearly (default daily)"
-// @Param        user_path   query     string  false  "Filter by tracked user path subtree"
-// @Param        cache_mode  query     string  false  "Cache mode filter: uncached, cached, all (default uncached)"
-// @Success      200  {array}   usage.DailyUsage
-// @Failure      400  {object}  core.GatewayError
-// @Failure      401  {object}  core.GatewayError
-// @Router       /admin/api/v1/usage/daily [get]
-func (h *Handler) DailyUsage(c *echo.Context) error {
-	return usageSliceResponse(c, h.usageReader, func(ctx context.Context, params usage.UsageQueryParams) ([]usage.DailyUsage, error) {
-		return h.usageReader.GetDailyUsage(ctx, params)
-	})
-}
-
-// UsageByModel handles GET /admin/api/v1/usage/models
-//
-// @Summary      Get usage breakdown by model
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Param        days        query     int     false  "Number of days (default 30)"
-// @Param        start_date  query     string  false  "Start date (YYYY-MM-DD)"
-// @Param        end_date    query     string  false  "End date (YYYY-MM-DD)"
-// @Param        user_path   query     string  false  "Filter by tracked user path subtree"
-// @Param        cache_mode  query     string  false  "Cache mode filter: uncached, cached, all (default uncached)"
-// @Success      200  {array}   usage.ModelUsage
-// @Failure      400  {object}  core.GatewayError
-// @Failure      401  {object}  core.GatewayError
-// @Router       /admin/api/v1/usage/models [get]
-func (h *Handler) UsageByModel(c *echo.Context) error {
-	return usageSliceResponse(c, h.usageReader, func(ctx context.Context, params usage.UsageQueryParams) ([]usage.ModelUsage, error) {
-		return h.usageReader.GetUsageByModel(ctx, params)
-	})
-}
-
-// UsageByUserPath handles GET /admin/api/v1/usage/user-paths
-//
-// @Summary      Get usage breakdown by user path
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Param        days        query     int     false  "Number of days (default 30)"
-// @Param        start_date  query     string  false  "Start date (YYYY-MM-DD)"
-// @Param        end_date    query     string  false  "End date (YYYY-MM-DD)"
-// @Param        user_path   query     string  false  "Filter by tracked user path subtree"
-// @Param        cache_mode  query     string  false  "Cache mode filter: uncached, cached, all (default uncached)"
-// @Success      200  {array}   usage.UserPathUsage
-// @Failure      400  {object}  core.GatewayError
-// @Failure      401  {object}  core.GatewayError
-// @Router       /admin/api/v1/usage/user-paths [get]
-func (h *Handler) UsageByUserPath(c *echo.Context) error {
-	return usageSliceResponse(c, h.usageReader, func(ctx context.Context, params usage.UsageQueryParams) ([]usage.UserPathUsage, error) {
-		return h.usageReader.GetUsageByUserPath(ctx, params)
-	})
-}
-
-// UsageLog handles GET /admin/api/v1/usage/log
-//
-// @Summary      Get paginated usage log entries
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Param        days        query     int     false  "Number of days (default 30)"
-// @Param        start_date  query     string  false  "Start date (YYYY-MM-DD)"
-// @Param        end_date    query     string  false  "End date (YYYY-MM-DD)"
-// @Param        model       query     string  false  "Filter by model name"
-// @Param        provider    query     string  false  "Filter by provider name or provider type"
-// @Param        user_path   query     string  false  "Filter by tracked user path subtree"
-// @Param        cache_mode  query     string  false  "Cache mode filter: uncached, cached, all (default uncached)"
-// @Param        search      query     string  false  "Search across model, provider, request_id, provider_id"
-// @Param        limit       query     int     false  "Page size (default 50, max 200)"
-// @Param        offset      query     int     false  "Offset for pagination"
-// @Success      200  {object}  usage.UsageLogResult
-// @Failure      400  {object}  core.GatewayError
-// @Failure      401  {object}  core.GatewayError
-// @Router       /admin/api/v1/usage/log [get]
-func (h *Handler) UsageLog(c *echo.Context) error {
-	if h.usageReader == nil {
-		return c.JSON(http.StatusOK, usage.UsageLogResult{
-			Entries: []usage.UsageLogEntry{},
-		})
-	}
-
-	baseParams, err := parseUsageParams(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	params := usage.UsageLogParams{
-		UsageQueryParams: baseParams,
-		Model:            c.QueryParam("model"),
-		Provider:         c.QueryParam("provider"),
-		Search:           c.QueryParam("search"),
-	}
-
-	if l := c.QueryParam("limit"); l != "" {
-		if parsed, err := strconv.Atoi(l); err == nil && parsed > 0 {
-			params.Limit = parsed
-		}
-	}
-	if o := c.QueryParam("offset"); o != "" {
-		if parsed, err := strconv.Atoi(o); err == nil && parsed >= 0 {
-			params.Offset = parsed
-		}
-	}
-
-	result, err := h.usageReader.GetUsageLog(c.Request().Context(), params)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	if result.Entries == nil {
-		result.Entries = []usage.UsageLogEntry{}
-	}
-
-	return c.JSON(http.StatusOK, result)
-}
-
-// RecalculateUsagePricing handles POST /admin/api/v1/usage/recalculate-pricing.
-//
-// @Summary      Recalculate stored usage costs from current model pricing metadata
-// @Tags         admin
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        request  body      recalculatePricingRequest       true  "Recalculation filters and confirmation"
-// @Success      200      {object}  usage.RecalculatePricingResult
-// @Failure      400      {object}  core.GatewayError
-// @Failure      401      {object}  core.GatewayError
-// @Failure      500      {object}  core.GatewayError
-// @Failure      503      {object}  core.GatewayError
-// @Router       /admin/api/v1/usage/recalculate-pricing [post]
-func (h *Handler) RecalculateUsagePricing(c *echo.Context) error {
-	if h.usageRecalculator == nil {
-		return handleError(c, featureUnavailableError("usage pricing recalculation is unavailable"))
-	}
-	if h.registry == nil {
-		return handleError(c, featureUnavailableError("model pricing metadata is unavailable"))
-	}
-
-	var req recalculatePricingRequest
-	if err := c.Bind(&req); err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-	if strings.TrimSpace(strings.ToLower(req.confirmationValue())) != "recalculate" {
-		return handleError(c, core.NewInvalidRequestError("confirmation must be recalculate", nil))
-	}
-
-	params, err := h.recalculatePricingParams(c, req)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	h.pricingMu.Lock()
-	defer h.pricingMu.Unlock()
-
-	result, err := h.usageRecalculator.RecalculatePricing(c.Request().Context(), params, h.registry)
-	if err != nil {
-		if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
-			return handleError(c, err)
-		}
-		if gatewayErr, ok := errors.AsType[*core.GatewayError](err); ok {
-			return handleError(c, gatewayErr)
-		}
-		return handleError(c, core.NewProviderError("usage", http.StatusInternalServerError, "failed to recalculate usage pricing", err))
-	}
-	return c.JSON(http.StatusOK, result)
-}
-
-// CacheOverview handles GET /admin/api/v1/cache/overview
-//
-// @Summary      Get cached-only usage overview
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Param        days        query     int     false  "Number of days (default 30)"
-// @Param        start_date  query     string  false  "Start date (YYYY-MM-DD)"
-// @Param        end_date    query     string  false  "End date (YYYY-MM-DD)"
-// @Param        interval    query     string  false  "Grouping interval: daily, weekly, monthly, yearly (default daily)"
-// @Param        user_path   query     string  false  "Filter by tracked user path subtree"
-// @Param        cache_mode  query     string  false  "Cache mode filter: uncached, cached, all (cache overview always uses cached mode)"
-// @Success      200  {object}  usage.CacheOverview
-// @Failure      400  {object}  core.GatewayError
-// @Failure      401  {object}  core.GatewayError
-// @Failure      503  {object}  core.GatewayError
-// @Router       /admin/api/v1/cache/overview [get]
-func (h *Handler) CacheOverview(c *echo.Context) error {
-	if strings.TrimSpace(h.runtimeConfig.CacheEnabled) != "on" {
-		return handleError(c, featureUnavailableError("cache analytics is unavailable"))
-	}
-	if h.usageReader == nil {
-		return c.JSON(http.StatusOK, usage.CacheOverview{
-			Daily: []usage.CacheOverviewDaily{},
-		})
-	}
-
-	params, err := parseUsageParams(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-	params.CacheMode = usage.CacheModeCached
-
-	overview, err := h.usageReader.GetCacheOverview(c.Request().Context(), params)
-	if err != nil {
-		return handleError(c, err)
-	}
-	if overview == nil {
-		overview = &usage.CacheOverview{}
-	}
-	if overview.Daily == nil {
-		overview.Daily = []usage.CacheOverviewDaily{}
-	}
-
-	return c.JSON(http.StatusOK, overview)
-}
-
-// AuditLog handles GET /admin/api/v1/audit/log
-//
-// @Summary      Get paginated audit log entries
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Param        days         query     int     false  "Number of days (default 30)"
-// @Param        start_date   query     string  false  "Start date (YYYY-MM-DD)"
-// @Param        end_date     query     string  false  "End date (YYYY-MM-DD)"
-// @Param        requested_model  query     string  false  "Filter by requested model selector"
-// @Param        provider     query     string  false  "Filter by provider name or provider type"
-// @Param        method       query     string  false  "Filter by HTTP method"
-// @Param        path         query     string  false  "Filter by request path"
-// @Param        user_path    query     string  false  "Filter by tracked user path subtree"
-// @Param        error_type   query     string  false  "Filter by error type"
-// @Param        status_code  query     int     false  "Filter by status code"
 // @Param        stream       query     bool    false  "Filter by stream mode (true/false)"
-// @Param        search       query     string  false  "Search across request_id/requested_model/provider/method/path/error_type/error_message"
-// @Param        limit        query     int     false  "Page size (default 25, max 100)"
-// @Param        offset       query     int     false  "Offset for pagination"
-// @Success      200  {object}  auditLogListResponse
-// @Failure      400  {object}  core.GatewayError
-// @Failure      401  {object}  core.GatewayError
-// @Router       /admin/api/v1/audit/log [get]
-func (h *Handler) AuditLog(c *echo.Context) error {
-	if h.auditReader == nil {
-		return c.JSON(http.StatusOK, auditLogListResponse{
-			Entries: []auditLogEntryResponse{},
-		})
-	}
-
-	dateRange, err := parseDateRangeParams(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-	userPath, err := normalizeUserPathQueryParam("user_path", c.QueryParam("user_path"))
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	requestedModel := c.QueryParam("requested_model")
-	if requestedModel == "" {
-		requestedModel = c.QueryParam("model")
-	}
-
-	params := auditlog.LogQueryParams{
-		QueryParams: auditlog.QueryParams{
-			StartDate: dateRange.StartDate,
-			EndDate:   dateRange.EndDate,
-		},
-		RequestedModel: requestedModel,
-		Provider:       c.QueryParam("provider"),
-		Method:         strings.ToUpper(c.QueryParam("method")),
-		Path:           c.QueryParam("path"),
-		UserPath:       userPath,
-		ErrorType:      c.QueryParam("error_type"),
-		Search:         c.QueryParam("search"),
-	}
-
-	if sc := c.QueryParam("status_code"); sc != "" {
-		parsed, err := strconv.Atoi(sc)
-		if err != nil {
-			return handleError(c, core.NewInvalidRequestError("invalid status_code, expected integer", nil))
-		}
-		params.StatusCode = &parsed
-	}
-
-	if stream := c.QueryParam("stream"); stream != "" {
-		parsed, err := strconv.ParseBool(stream)
-		if err != nil {
-			return handleError(c, core.NewInvalidRequestError("invalid stream value, expected true or false", nil))
-		}
-		params.Stream = &parsed
-	}
-
-	if l := c.QueryParam("limit"); l != "" {
-		if parsed, err := strconv.Atoi(l); err == nil && parsed > 0 {
-			params.Limit = parsed
-		}
-	}
-	if o := c.QueryParam("offset"); o != "" {
-		if parsed, err := strconv.Atoi(o); err == nil && parsed >= 0 {
-			params.Offset = parsed
-		}
-	}
-
-	result, err := h.auditReader.GetLogs(c.Request().Context(), params)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	if result.Entries == nil {
-		result.Entries = []auditlog.LogEntry{}
-	}
-
-	response, err := h.auditLogResponse(c.Request().Context(), result)
-	if err != nil {
-		return handleError(c, err)
-	}
-	return c.JSON(http.StatusOK, response)
-}
-
-func (h *Handler) auditLogResponse(ctx context.Context, result *auditlog.LogListResult) (*auditLogListResponse, error) {
-	if result == nil {
-		return &auditLogListResponse{Entries: []auditLogEntryResponse{}}, nil
-	}
-
-	response := &auditLogListResponse{
-		Entries: make([]auditLogEntryResponse, len(result.Entries)),
-		Total:   result.Total,
-		Limit:   result.Limit,
-		Offset:  result.Offset,
-	}
-	for i := range result.Entries {
-		response.Entries[i].LogEntry = result.Entries[i]
-	}
-
-	if h.usageReader == nil || len(result.Entries) == 0 {
-		return response, nil
-	}
-
-	requestIDs := make([]string, 0, len(result.Entries))
-	for _, entry := range result.Entries {
-		requestIDs = append(requestIDs, entry.RequestID)
-	}
-
-	entriesByRequestID, err := h.usageReader.GetUsageByRequestIDs(ctx, requestIDs)
-	if err != nil {
-		slog.Warn("failed to enrich audit log entries with usage", "error", err, "request_count", len(requestIDs))
-		return response, nil
-	}
-
-	summaries := usage.SummarizeUsageByRequestID(entriesByRequestID)
-	for i := range response.Entries {
-		requestID := response.Entries[i].RequestID
-		if summary, ok := summaries[requestID]; ok {
-			response.Entries[i].Usage = summary
-		}
-	}
-
-	return response, nil
-}
-
-// AuditConversation handles GET /admin/api/v1/audit/conversation
-//
-// @Summary      Get conversation thread around an audit log entry
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Param        log_id  query     string  true   "Anchor audit log entry ID"
-// @Param        limit   query     int     false  "Max entries in thread (default 40, max 200)"
-// @Success      200  {object}  auditlog.ConversationResult
-// @Failure      400  {object}  core.GatewayError
-// @Failure      401  {object}  core.GatewayError
-// @Router       /admin/api/v1/audit/conversation [get]
-func (h *Handler) AuditConversation(c *echo.Context) error {
-	if h.auditReader == nil {
-		return c.JSON(http.StatusOK, auditlog.ConversationResult{
-			AnchorID: c.QueryParam("log_id"),
-			Entries:  []auditlog.LogEntry{},
-		})
-	}
-
-	logID := strings.TrimSpace(c.QueryParam("log_id"))
-	if logID == "" {
-		return handleError(c, core.NewInvalidRequestError("log_id is required", nil))
-	}
-
-	limit := 40
-	if l := c.QueryParam("limit"); l != "" {
-		parsed, err := strconv.Atoi(l)
-		if err != nil {
-			return handleError(c, core.NewInvalidRequestError("invalid limit, expected integer", nil))
-		}
-		if parsed < 1 || parsed > 200 {
-			return handleError(c, core.NewInvalidRequestError("invalid limit parameter: limit must be between 1 and 200", nil))
-		}
-		limit = parsed
-	}
-
-	result, err := h.auditReader.GetConversation(c.Request().Context(), logID, limit)
-	if err != nil {
-		return handleError(c, err)
-	}
-	if result == nil {
-		result = &auditlog.ConversationResult{
-			AnchorID: logID,
-			Entries:  []auditlog.LogEntry{},
-		}
-	}
-	if result.Entries == nil {
-		result.Entries = []auditlog.LogEntry{}
-	}
-
-	return c.JSON(http.StatusOK, result)
-}
-
-// ListModels handles GET /admin/api/v1/models
-// Supports optional ?category= query param for filtering by model category.
-//
-// @Summary      List all registered models with provider info
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Success      200  {array}  providers.ModelWithProvider
-// @Failure      401  {object}  core.GatewayError
 // @Router       /admin/api/v1/models [get]
-type modelAccessResponse struct {
-	Selector         string                   `json:"selector"`
-	DefaultEnabled   bool                     `json:"default_enabled"`
-	EffectiveEnabled bool                     `json:"effective_enabled"`
-	UserPaths        []string                 `json:"user_paths,omitempty"`
-	Override         *modeloverrides.Override `json:"override,omitempty"`
-}
-
-type modelInventoryResponse struct {
-	providers.ModelWithProvider
-	Access modelAccessResponse `json:"access"`
-}
-
-func (h *Handler) ListModels(c *echo.Context) error {
-	if h.registry == nil {
-		return c.JSON(http.StatusOK, []modelInventoryResponse{})
-	}
-
-	cat := core.ModelCategory(c.QueryParam("category"))
-	if cat != "" && cat != core.CategoryAll {
-		if !isValidCategory(cat) {
-			return handleError(c, core.NewInvalidRequestError("invalid category: "+string(cat), nil))
-		}
-	}
-
-	var models []providers.ModelWithProvider
-	if cat != "" && cat != core.CategoryAll {
-		models = h.registry.ListModelsWithProviderByCategory(cat)
-	} else {
-		models = h.registry.ListModelsWithProvider()
-	}
-
-	if models == nil {
-		models = []providers.ModelWithProvider{}
-	}
-	if h.modelOverrides == nil {
-		response := make([]modelInventoryResponse, 0, len(models))
-		for _, model := range models {
-			selector := core.ModelSelector{
-				Provider: strings.TrimSpace(model.ProviderName),
-				Model:    strings.TrimSpace(model.Model.ID),
-			}
-			response = append(response, modelInventoryResponse{
-				ModelWithProvider: model,
-				Access: modelAccessResponse{
-					Selector:         selector.QualifiedModel(),
-					DefaultEnabled:   true,
-					EffectiveEnabled: true,
-				},
-			})
-		}
-		return c.JSON(http.StatusOK, response)
-	}
-
-	response := make([]modelInventoryResponse, 0, len(models))
-	for _, model := range models {
-		selector := core.ModelSelector{
-			Provider: strings.TrimSpace(model.ProviderName),
-			Model:    strings.TrimSpace(model.Model.ID),
-		}
-		effective := h.modelOverrides.EffectiveState(selector)
-		access := modelAccessResponse{
-			Selector:         effective.Selector,
-			DefaultEnabled:   effective.DefaultEnabled,
-			EffectiveEnabled: effective.Enabled,
-			UserPaths:        append([]string(nil), effective.UserPaths...),
-		}
-		if override, ok := h.modelOverrides.Get(selector.QualifiedModel()); ok && override != nil {
-			overrideCopy := *override
-			access.Override = &overrideCopy
-		}
-		response = append(response, modelInventoryResponse{
-			ModelWithProvider: model,
-			Access:            access,
-		})
-	}
-
-	return c.JSON(http.StatusOK, response)
-}
-
-// isValidCategory returns true if cat is a recognized model category.
-func isValidCategory(cat core.ModelCategory) bool {
-	return slices.Contains(core.AllCategories(), cat)
-}
-
-// ListCategories handles GET /admin/api/v1/models/categories
-//
-// @Summary      List model categories with counts
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Success      200  {array}   providers.CategoryCount
-// @Failure      401  {object}  core.GatewayError
-// @Router       /admin/api/v1/models/categories [get]
-func (h *Handler) ListCategories(c *echo.Context) error {
-	if h.registry == nil {
-		return c.JSON(http.StatusOK, []providers.CategoryCount{})
-	}
-
-	return c.JSON(http.StatusOK, h.registry.GetCategoryCounts())
-}
-
-// DashboardConfig handles GET /admin/api/v1/dashboard/config
-func (h *Handler) DashboardConfig(c *echo.Context) error {
-	return c.JSON(http.StatusOK, cloneDashboardRuntimeConfig(h.runtimeConfig))
-}
-
-// ListBudgets handles GET /admin/api/v1/budgets.
-// @Summary      List budgets with current status
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Success      200  {object}  budgetListResponse
-// @Failure      401  {object}  core.GatewayError
-// @Failure      503  {object}  core.GatewayError
 // @Router       /admin/api/v1/budgets [get]
-func (h *Handler) ListBudgets(c *echo.Context) error {
-	if h.budgets == nil {
-		return handleError(c, featureUnavailableError("budgets feature is unavailable"))
-	}
-	now := time.Now().UTC()
-	statuses, err := h.budgets.Statuses(c.Request().Context(), now)
-	if err != nil {
-		return handleError(c, budgetServiceError("failed to list budgets", err))
-	}
-	return c.JSON(http.StatusOK, budgetListResponse{
-		Budgets:    budgetStatusResponses(statuses, now),
-		ServerTime: now,
-	})
-}
-
-// UpsertBudget handles PUT /admin/api/v1/budgets/{user_path}/{period}.
-// @Summary      Create or update one budget
-// @Tags         admin
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        user_path  path      string               true  "URL-encoded budget user path"
-// @Param        period     path      string               true  "Budget period name or seconds"
-// @Param        budget     body      upsertBudgetRequest  true  "Budget amount"
-// @Success      200        {object}  budgetListResponse
-// @Failure      400        {object}  core.GatewayError
-// @Failure      401        {object}  core.GatewayError
-// @Failure      503        {object}  core.GatewayError
-// @Router       /admin/api/v1/budgets/{user_path}/{period} [put]
-func (h *Handler) UpsertBudget(c *echo.Context) error {
-	if h.budgets == nil {
-		return handleError(c, featureUnavailableError("budgets feature is unavailable"))
-	}
-	var req upsertBudgetRequest
-	if err := c.Bind(&req); err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-	userPath, periodSeconds, err := budgetRouteKey(c)
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError(err.Error(), err))
-	}
-	item, err := budget.NormalizeBudget(budget.Budget{
-		UserPath:      userPath,
-		PeriodSeconds: periodSeconds,
-		Amount:        req.Amount,
-		Source:        budget.SourceManual,
-	})
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError(err.Error(), err))
-	}
-	if err := h.budgets.UpsertBudgets(c.Request().Context(), []budget.Budget{item}); err != nil {
-		return handleError(c, budgetServiceError("failed to save budget", err))
-	}
-	return h.ListBudgets(c)
-}
-
-// DeleteBudget handles DELETE /admin/api/v1/budgets/{user_path}/{period}.
-// @Summary      Delete one budget
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Param        user_path  path      string  true  "URL-encoded budget user path"
-// @Param        period     path      string  true  "Budget period name or seconds"
-// @Success      200        {object}  budgetListResponse
-// @Failure      400        {object}  core.GatewayError
-// @Failure      401        {object}  core.GatewayError
-// @Failure      503        {object}  core.GatewayError
-// @Router       /admin/api/v1/budgets/{user_path}/{period} [delete]
-func (h *Handler) DeleteBudget(c *echo.Context) error {
-	if h.budgets == nil {
-		return handleError(c, featureUnavailableError("budgets feature is unavailable"))
-	}
-	userPath, periodSeconds, err := budgetRouteKey(c)
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError(err.Error(), err))
-	}
-	if err := h.budgets.DeleteBudget(c.Request().Context(), userPath, periodSeconds); err != nil {
-		return handleError(c, budgetServiceError("failed to delete budget", err))
-	}
-	return h.ListBudgets(c)
-}
-
-// BudgetSettings handles GET /admin/api/v1/budgets/settings.
-// @Summary      Get budget reset settings
-// @Tags         admin
-// @Produce      json
-// @Security     BearerAuth
-// @Success      200  {object}  budget.Settings
-// @Failure      401  {object}  core.GatewayError
-// @Failure      503  {object}  core.GatewayError
-// @Router       /admin/api/v1/budgets/settings [get]
-func (h *Handler) BudgetSettings(c *echo.Context) error {
-	if h.budgets == nil {
-		return handleError(c, featureUnavailableError("budgets feature is unavailable"))
-	}
-	return c.JSON(http.StatusOK, h.budgets.Settings())
-}
-
-// UpdateBudgetSettings handles PUT /admin/api/v1/budgets/settings.
-// @Summary      Update budget reset settings
-// @Tags         admin
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        settings  body      updateBudgetSettingsRequest  true  "Budget reset settings"
-// @Success      200       {object}  budget.Settings
-// @Failure      400       {object}  core.GatewayError
-// @Failure      401       {object}  core.GatewayError
-// @Failure      503       {object}  core.GatewayError
-// @Router       /admin/api/v1/budgets/settings [put]
-func (h *Handler) UpdateBudgetSettings(c *echo.Context) error {
-	if h.budgets == nil {
-		return handleError(c, featureUnavailableError("budgets feature is unavailable"))
-	}
-	var req updateBudgetSettingsRequest
-	if err := c.Bind(&req); err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-	settings := req.apply(h.budgets.Settings())
-	if err := budget.ValidateSettings(settings); err != nil {
-		return handleError(c, core.NewInvalidRequestError(err.Error(), err))
-	}
-	saved, err := h.budgets.SaveSettings(c.Request().Context(), settings)
-	if err != nil {
-		return handleError(c, budgetServiceError("failed to save budget settings", err))
-	}
-	return c.JSON(http.StatusOK, saved)
-}
-
-// ResetBudget handles POST /admin/api/v1/budgets/reset-one.
-// @Summary      Reset one budget period
-// @Tags         admin
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        budget  body      resetBudgetRequest  true  "Budget key"
-// @Success      200     {object}  budgetListResponse
-// @Failure      400     {object}  core.GatewayError
-// @Failure      401     {object}  core.GatewayError
-// @Failure      503     {object}  core.GatewayError
-// @Router       /admin/api/v1/budgets/reset-one [post]
-func (h *Handler) ResetBudget(c *echo.Context) error {
-	if h.budgets == nil {
-		return handleError(c, featureUnavailableError("budgets feature is unavailable"))
-	}
-	var req resetBudgetRequest
-	if err := c.Bind(&req); err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-	periodSeconds, err := budgetRequestPeriodSeconds(req.Period, req.PeriodSeconds)
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError(err.Error(), err))
-	}
-	userPath, err := budget.NormalizeUserPath(req.UserPath)
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError(err.Error(), err))
-	}
-	if err := h.budgets.ResetBudget(c.Request().Context(), userPath, periodSeconds, time.Now().UTC()); err != nil {
-		return handleError(c, budgetServiceError("failed to reset budget", err))
-	}
-	return h.ListBudgets(c)
-}
-
-// ResetBudgets handles POST /admin/api/v1/budgets/reset.
-// @Summary      Reset all budget periods
-// @Tags         admin
-// @Accept       json
-// @Produce      json
-// @Security     BearerAuth
-// @Param        confirmation  body      resetBudgetsRequest  true  "Reset confirmation"
-// @Success      200           {object}  resetBudgetsResponse
-// @Failure      400           {object}  core.GatewayError
-// @Failure      401           {object}  core.GatewayError
-// @Failure      503           {object}  core.GatewayError
-// @Router       /admin/api/v1/budgets/reset [post]
-func (h *Handler) ResetBudgets(c *echo.Context) error {
-	if h.budgets == nil {
-		return handleError(c, featureUnavailableError("budgets feature is unavailable"))
-	}
-	var req resetBudgetsRequest
-	if err := c.Bind(&req); err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-	if strings.TrimSpace(strings.ToLower(req.confirmationValue())) != "reset" {
-		return handleError(c, core.NewInvalidRequestError("confirmation must be reset", nil))
-	}
-	if err := h.budgets.ResetAll(c.Request().Context(), time.Now().UTC()); err != nil {
-		return handleError(c, budgetServiceError("failed to reset budgets", err))
-	}
-	return c.JSON(http.StatusOK, resetBudgetsResponse{Status: "ok"})
-}
-
-// ProviderStatus handles GET /admin/api/v1/providers/status
-func (h *Handler) ProviderStatus(c *echo.Context) error {
-	return c.JSON(http.StatusOK, h.buildProviderStatusResponse())
-}
-
-// RefreshRuntime handles POST /admin/api/v1/runtime/refresh
-func (h *Handler) RefreshRuntime(c *echo.Context) error {
-	if h.runtimeRefresher == nil {
-		return handleError(c, featureUnavailableError("runtime refresh is unavailable"))
-	}
-
-	report, err := h.runtimeRefresher.RefreshRuntime(c.Request().Context())
-	if err != nil {
-		if gatewayErr, ok := errors.AsType[*core.GatewayError](err); ok {
-			return handleError(c, gatewayErr)
-		}
-		return handleError(c, core.NewProviderError("runtime_refresh", http.StatusInternalServerError, "runtime refresh failed", err))
-	}
-	if report.Status == "" {
-		report.Status = RuntimeRefreshStatusOK
-	}
-	if report.Steps == nil {
-		report.Steps = []RuntimeRefreshStep{}
-	}
-	return c.JSON(http.StatusOK, report)
-}
-
-func (h *Handler) buildProviderStatusResponse() providerStatusResponse {
-	configured := cloneConfiguredProviders(h.configuredProviders)
-	configuredByName := make(map[string]providers.SanitizedProviderConfig, len(configured))
-	nameSet := make(map[string]struct{}, len(configured))
-	for _, cfg := range configured {
-		name := strings.TrimSpace(cfg.Name)
-		if name == "" {
-			continue
-		}
-		configuredByName[name] = cfg
-		nameSet[name] = struct{}{}
-	}
-
-	runtimeByName := make(map[string]providers.ProviderRuntimeSnapshot)
-	if h.registry != nil {
-		for _, snapshot := range h.registry.ProviderRuntimeSnapshots() {
-			name := strings.TrimSpace(snapshot.Name)
-			if name == "" {
-				continue
-			}
-			runtimeByName[name] = snapshot
-			nameSet[name] = struct{}{}
-		}
-	}
-
-	names := make([]string, 0, len(nameSet))
-	for name := range nameSet {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-
-	resp := providerStatusResponse{
-		Summary: providerStatusSummaryResponse{
-			OverallStatus: "degraded",
-		},
-		Providers: make([]providerStatusItemResponse, 0, len(names)),
-	}
-
-	for _, name := range names {
-		cfg, hasConfig := configuredByName[name]
-		runtime, hasRuntime := runtimeByName[name]
-		if !hasConfig {
-			cfg = providers.SanitizedProviderConfig{Name: name, Type: strings.TrimSpace(runtime.Type)}
-		}
-		if !hasRuntime {
-			runtime = providers.ProviderRuntimeSnapshot{Name: name, Type: strings.TrimSpace(cfg.Type)}
-		}
-		if strings.TrimSpace(cfg.Type) == "" {
-			cfg.Type = strings.TrimSpace(runtime.Type)
-		}
-		if strings.TrimSpace(runtime.Type) == "" {
-			runtime.Type = strings.TrimSpace(cfg.Type)
-		}
-
-		status, label, reason, lastError := classifyProviderStatus(cfg, runtime)
-		resp.Providers = append(resp.Providers, providerStatusItemResponse{
-			Name:         name,
-			Type:         strings.TrimSpace(cfg.Type),
-			Status:       status,
-			StatusLabel:  label,
-			StatusReason: reason,
-			LastError:    lastError,
-			Config:       cfg,
-			Runtime:      runtime,
-		})
-		resp.Summary.Total++
-		switch status {
-		case "healthy":
-			resp.Summary.Healthy++
-		case "unhealthy":
-			resp.Summary.Unhealthy++
-		default:
-			resp.Summary.Degraded++
-		}
-	}
-
-	switch {
-	case resp.Summary.Total == 0:
-		resp.Summary.OverallStatus = "degraded"
-	case resp.Summary.Healthy == resp.Summary.Total:
-		resp.Summary.OverallStatus = "healthy"
-	case resp.Summary.Unhealthy == resp.Summary.Total:
-		resp.Summary.OverallStatus = "unhealthy"
-	default:
-		resp.Summary.OverallStatus = "degraded"
-	}
-
-	if resp.Providers == nil {
-		resp.Providers = []providerStatusItemResponse{}
-	}
-	return resp
-}
-
-func classifyProviderStatus(cfg providers.SanitizedProviderConfig, runtime providers.ProviderRuntimeSnapshot) (status, label, reason, lastError string) {
-	modelFetchError := strings.TrimSpace(runtime.LastModelFetchError)
-	availabilityError := strings.TrimSpace(runtime.LastAvailabilityError)
-	configuredName := strings.TrimSpace(cfg.Name)
-	usingCachedModels := runtime.Registered &&
-		runtime.DiscoveredModelCount > 0 &&
-		modelFetchError == "" &&
-		runtime.LastModelFetchSuccessAt == nil
-
-	lastError = modelFetchError
-	if lastError == "" {
-		lastError = availabilityError
-	}
-
-	switch {
-	case runtime.DiscoveredModelCount > 0 && modelFetchError == "":
-		if usingCachedModels {
-			return "degraded", "Starting", "serving cached model inventory while live refresh finishes", lastError
-		}
-		return "healthy", "Healthy", "configured and model discovery succeeded", lastError
-	case modelFetchError != "" && runtime.DiscoveredModelCount > 0:
-		return "degraded", "Degraded", "latest model refresh failed; previous inventory is still available", lastError
-	case modelFetchError != "":
-		return "unhealthy", "Unhealthy", "model discovery failed and no provider models are currently available", lastError
-	case availabilityError != "" && runtime.DiscoveredModelCount == 0:
-		return "unhealthy", "Unhealthy", "startup availability check failed and no provider models are available", lastError
-	case runtime.DiscoveredModelCount > 0:
-		return "healthy", "Healthy", "provider models are currently available", lastError
-	case !runtime.Registered && configuredName != "":
-		return "degraded", "Starting", "provider is configured and awaiting live model discovery", lastError
-	case configuredName != "":
-		return "degraded", "Configured", "provider is configured but has not exposed models yet", lastError
-	default:
-		return "degraded", "Unknown", "provider runtime inventory is unavailable", lastError
-	}
-}
-
-type upsertAliasRequest struct {
-	TargetModel    string `json:"target_model"`
-	TargetProvider string `json:"target_provider,omitempty"`
-	Description    string `json:"description,omitempty"`
-	Enabled        *bool  `json:"enabled,omitempty"`
-}
-
-type upsertModelOverrideRequest struct {
-	UserPaths []string `json:"user_paths,omitempty"`
-}
-
-type upsertGuardrailRequest struct {
-	Type        string          `json:"type"`
-	Description string          `json:"description,omitempty"`
-	UserPath    string          `json:"user_path,omitempty"`
-	Config      json.RawMessage `json:"config"`
-}
-
-type createWorkflowRequest struct {
-	ScopeProviderName   string            `json:"scope_provider_name,omitempty"`
-	LegacyScopeProvider string            `json:"scope_provider,omitempty"`
-	ScopeModel          string            `json:"scope_model,omitempty"`
-	ScopeUserPath       string            `json:"scope_user_path,omitempty"`
-	Name                string            `json:"name"`
-	Description         string            `json:"description,omitempty"`
-	Payload             workflows.Payload `json:"workflow_payload"`
-}
-
-type createAuthKeyRequest struct {
-	Name        string     `json:"name"`
-	Description string     `json:"description,omitempty"`
-	UserPath    string     `json:"user_path,omitempty"`
-	ExpiresAt   *time.Time `json:"expires_at,omitempty"`
-}
-
-type budgetListResponse struct {
-	Budgets    []budgetStatusResponse `json:"budgets"`
-	ServerTime time.Time              `json:"server_time"`
-}
-
-type budgetStatusResponse struct {
-	UserPath      string     `json:"user_path"`
-	PeriodSeconds int64      `json:"period_seconds"`
-	PeriodLabel   string     `json:"period_label"`
-	Amount        float64    `json:"amount"`
-	Source        string     `json:"source,omitempty"`
-	LastResetAt   *time.Time `json:"last_reset_at,omitempty"`
-	CreatedAt     time.Time  `json:"created_at"`
-	UpdatedAt     time.Time  `json:"updated_at"`
-	PeriodStart   time.Time  `json:"period_start"`
-	PeriodEnd     time.Time  `json:"period_end"`
-	Spent         float64    `json:"spent"`
-	HasUsage      bool       `json:"has_usage"`
-	Remaining     float64    `json:"remaining"`
-	UsageRatio    float64    `json:"usage_ratio"`
-	PeriodRatio   float64    `json:"period_ratio"`
-}
-
-type upsertBudgetRequest struct {
-	Amount float64 `json:"amount"`
-}
-
-type resetBudgetRequest struct {
-	UserPath      string `json:"user_path"`
-	Period        string `json:"period,omitempty"`
-	PeriodSeconds int64  `json:"period_seconds,omitempty"`
-}
-
-type updateBudgetSettingsRequest struct {
-	DailyResetHour     *int `json:"daily_reset_hour"`
-	DailyResetMinute   *int `json:"daily_reset_minute"`
-	WeeklyResetWeekday *int `json:"weekly_reset_weekday"`
-	WeeklyResetHour    *int `json:"weekly_reset_hour"`
-	WeeklyResetMinute  *int `json:"weekly_reset_minute"`
-	MonthlyResetDay    *int `json:"monthly_reset_day"`
-	MonthlyResetHour   *int `json:"monthly_reset_hour"`
-	MonthlyResetMinute *int `json:"monthly_reset_minute"`
-}
-
-type recalculatePricingRequest struct {
-	Days         int    `json:"days,omitempty"`
-	StartDate    string `json:"start_date,omitempty"`
-	EndDate      string `json:"end_date,omitempty"`
-	UserPath     string `json:"user_path,omitempty"`
-	Selector     string `json:"selector,omitempty"`
-	Confirmation string `json:"confirmation"`
-	Confirm      string `json:"confirm,omitempty"`
-}
-
-func (r recalculatePricingRequest) confirmationValue() string {
-	if strings.TrimSpace(r.Confirmation) != "" {
-		return r.Confirmation
-	}
-	return r.Confirm
-}
-
-func (r updateBudgetSettingsRequest) apply(settings budget.Settings) budget.Settings {
-	if r.DailyResetHour != nil {
-		settings.DailyResetHour = *r.DailyResetHour
-	}
-	if r.DailyResetMinute != nil {
-		settings.DailyResetMinute = *r.DailyResetMinute
-	}
-	if r.WeeklyResetWeekday != nil {
-		settings.WeeklyResetWeekday = *r.WeeklyResetWeekday
-	}
-	if r.WeeklyResetHour != nil {
-		settings.WeeklyResetHour = *r.WeeklyResetHour
-	}
-	if r.WeeklyResetMinute != nil {
-		settings.WeeklyResetMinute = *r.WeeklyResetMinute
-	}
-	if r.MonthlyResetDay != nil {
-		settings.MonthlyResetDay = *r.MonthlyResetDay
-	}
-	if r.MonthlyResetHour != nil {
-		settings.MonthlyResetHour = *r.MonthlyResetHour
-	}
-	if r.MonthlyResetMinute != nil {
-		settings.MonthlyResetMinute = *r.MonthlyResetMinute
-	}
-	return settings
-}
-
-type resetBudgetsRequest struct {
-	Confirmation string `json:"confirmation"`
-	Confirm      string `json:"confirm,omitempty"`
-}
-
-func (r resetBudgetsRequest) confirmationValue() string {
-	if strings.TrimSpace(r.Confirmation) != "" {
-		return r.Confirmation
-	}
-	return r.Confirm
-}
-
-type resetBudgetsResponse struct {
-	Status string `json:"status"`
-}
-
-func (h *Handler) recalculatePricingParams(c *echo.Context, req recalculatePricingRequest) (usage.RecalculatePricingParams, error) {
-	baseParams, err := recalculatePricingDateParams(c, req)
-	if err != nil {
-		return usage.RecalculatePricingParams{}, err
-	}
-
-	userPath, err := normalizeUserPathQueryParam("user_path", req.UserPath)
-	if err != nil {
-		return usage.RecalculatePricingParams{}, err
-	}
-	baseParams.UserPath = userPath
-	baseParams.CacheMode = usage.CacheModeAll
-
-	provider, model, err := h.recalculatePricingSelector(req.Selector)
-	if err != nil {
-		return usage.RecalculatePricingParams{}, err
-	}
-
-	return usage.RecalculatePricingParams{
-		UsageQueryParams: baseParams,
-		Provider:         provider,
-		Model:            model,
-	}, nil
-}
-
-func recalculatePricingDateParams(c *echo.Context, req recalculatePricingRequest) (usage.UsageQueryParams, error) {
-	var params usage.UsageQueryParams
-
-	timeZone, location := dashboardTimeZone(c)
-	params.TimeZone = timeZone
-
-	now := timeNow().In(location)
-	today := time.Date(now.Year(), now.Month(), now.Day(), 0, 0, 0, 0, location)
-
-	start, end, err := buildDateRange(strings.TrimSpace(req.StartDate), strings.TrimSpace(req.EndDate), normalizeDateRangeDays(req.Days), location, today)
-	if err != nil {
-		return params, err
-	}
-	params.StartDate = start
-	params.EndDate = end
-	return params, nil
-}
-
-func (h *Handler) recalculatePricingSelector(raw string) (provider, model string, err error) {
-	raw = strings.TrimSpace(raw)
-	if raw == "" {
-		return "", "", nil
-	}
-
-	if h.aliases != nil {
-		selector, changed, err := h.aliases.ResolveModel(core.NewRequestedModelSelector(raw, ""))
-		if err != nil {
-			return "", "", core.NewInvalidRequestError("invalid selector: "+err.Error(), err)
-		}
-		if changed {
-			return selector.Provider, selector.Model, nil
-		}
-	}
-
-	selector, err := core.ParseModelSelector(raw, "")
-	if err != nil {
-		return "", "", core.NewInvalidRequestError("invalid selector: "+err.Error(), err)
-	}
-	if selector.Provider == "" {
-		return "", "", core.NewInvalidRequestError("invalid selector: provider/model or alias is required", nil)
-	}
-	return selector.Provider, selector.Model, nil
-}
-
-func budgetStatusResponses(statuses []budget.CheckResult, now time.Time) []budgetStatusResponse {
-	if len(statuses) == 0 {
-		return []budgetStatusResponse{}
-	}
-	responses := make([]budgetStatusResponse, 0, len(statuses))
-	for _, status := range statuses {
-		item := status.Budget
-		usageRatio := 0.0
-		if item.Amount > 0 {
-			usageRatio = status.Spent / item.Amount
-		}
-		periodRatio := 0.0
-		periodDuration := status.PeriodEnd.Sub(status.PeriodStart).Seconds()
-		if periodDuration > 0 {
-			periodRatio = now.Sub(status.PeriodStart).Seconds() / periodDuration
-		}
-		responses = append(responses, budgetStatusResponse{
-			UserPath:      item.UserPath,
-			PeriodSeconds: item.PeriodSeconds,
-			PeriodLabel:   budget.PeriodLabel(item.PeriodSeconds),
-			Amount:        item.Amount,
-			Source:        item.Source,
-			LastResetAt:   item.LastResetAt,
-			CreatedAt:     item.CreatedAt,
-			UpdatedAt:     item.UpdatedAt,
-			PeriodStart:   status.PeriodStart,
-			PeriodEnd:     status.PeriodEnd,
-			Spent:         status.Spent,
-			HasUsage:      status.HasUsage,
-			Remaining:     status.Remaining,
-			UsageRatio:    usageRatio,
-			PeriodRatio:   clampBudgetRatio(periodRatio),
-		})
-	}
-	return responses
-}
-
-func budgetRouteKey(c *echo.Context) (string, int64, error) {
-	userPathParam := strings.TrimSpace(c.Param("user_path"))
-	if userPathParam == "" {
-		return "", 0, errors.New("user_path path parameter is required")
-	}
-	userPath, err := url.PathUnescape(userPathParam)
-	if err != nil {
-		return "", 0, fmt.Errorf("invalid user_path path parameter: %w", err)
-	}
-	userPath, err = budget.NormalizeUserPath(userPath)
-	if err != nil {
-		return "", 0, err
-	}
-
-	periodParam := strings.TrimSpace(c.Param("period"))
-	if periodParam == "" {
-		return "", 0, errors.New("period path parameter is required")
-	}
-	if seconds, err := strconv.ParseInt(periodParam, 10, 64); err == nil {
-		if seconds <= 0 {
-			return "", 0, errors.New("period_seconds must be greater than 0")
-		}
-		return userPath, seconds, nil
-	}
-	periodSeconds, err := budgetRequestPeriodSeconds(periodParam, 0)
-	if err != nil {
-		return "", 0, err
-	}
-	return userPath, periodSeconds, nil
-}
-
-func budgetRequestPeriodSeconds(period string, periodSeconds int64) (int64, error) {
-	if periodSeconds > 0 {
-		return periodSeconds, nil
-	}
-	if parsed, ok := budget.PeriodSeconds(period); ok {
-		return parsed, nil
-	}
-	return 0, errors.New("period must be one of hourly, daily, weekly, monthly or period_seconds must be set")
-}
-
-func clampBudgetRatio(value float64) float64 {
-	if value < 0 {
-		return 0
-	}
-	if value > 1 {
-		return 1
-	}
-	return value
-}
-
-func budgetServiceError(message string, err error) error {
-	if errors.Is(err, budget.ErrNotFound) {
-		return core.NewNotFoundError("budget not found").WithCode("budget_not_found")
-	}
-	return core.NewProviderError("budgets", http.StatusServiceUnavailable, message, err)
-}
-
-func featureUnavailableError(message string) error {
-	return core.NewInvalidRequestErrorWithStatus(http.StatusServiceUnavailable, message, nil).
-		WithCode("feature_unavailable")
-}
-
-func validationWriter(isValidation func(error) bool) func(error) error {
-	return func(err error) error {
-		if err == nil {
-			return nil
-		}
-		if isValidation(err) {
-			return core.NewInvalidRequestError(err.Error(), err)
-		}
-		return err
-	}
-}
-
-var (
-	aliasWriteError     = validationWriter(aliases.IsValidationError)
-	workflowWriteError  = validationWriter(workflows.IsValidationError)
-	authKeyWriteError   = validationWriter(authkeys.IsValidationError)
-	guardrailWriteError = validationWriter(guardrails.IsValidationError)
-)
-
-// modelOverrideWriteError differs from the others: non-validation errors are
-// surfaced as 502 so the dashboard distinguishes provider failures from input issues.
-func modelOverrideWriteError(err error) error {
-	if err == nil {
-		return nil
-	}
-	if modeloverrides.IsValidationError(err) {
-		return core.NewInvalidRequestError(err.Error(), err)
-	}
-	return core.NewProviderError("model_overrides", http.StatusBadGateway, err.Error(), err)
-}
-
-func deactivateByID(
-	c *echo.Context,
-	unavailableErr error,
-	idLabel string,
-	notFoundErr error,
-	notFoundMessage string,
-	deactivate func(context.Context, string) error,
-	writeError func(error) error,
-) error {
-	if unavailableErr != nil {
-		return handleError(c, unavailableErr)
-	}
-
-	id := strings.TrimSpace(c.Param("id"))
-	if id == "" {
-		return handleError(c, core.NewInvalidRequestError(idLabel+" id is required", nil))
-	}
-
-	if err := deactivate(c.Request().Context(), id); err != nil {
-		if errors.Is(err, notFoundErr) {
-			return handleError(c, core.NewNotFoundError(notFoundMessage+id))
-		}
-		return handleError(c, writeError(err))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-func deleteByName(
-	c *echo.Context,
-	unavailableErr error,
-	paramName string,
-	decode func(string) (string, error),
-	deleteFunc func(context.Context, string) error,
-	notFoundErr error,
-	notFoundMessage string,
-	writeError func(error) error,
-) error {
-	if unavailableErr != nil {
-		return handleError(c, unavailableErr)
-	}
-
-	name, err := decode(c.Param(paramName))
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	if err := deleteFunc(c.Request().Context(), name); err != nil {
-		if errors.Is(err, notFoundErr) {
-			return handleError(c, core.NewNotFoundError(notFoundMessage+name))
-		}
-		return handleError(c, writeError(err))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
 
 // ListModelOverrides handles GET /admin/api/v1/model-overrides.
-func (h *Handler) ListModelOverrides(c *echo.Context) error {
-	if h.modelOverrides == nil {
-		return handleError(c, featureUnavailableError("model overrides feature is unavailable"))
-	}
-	views := h.modelOverrides.ListViews()
-	if views == nil {
-		views = []modeloverrides.View{}
-	}
-	return c.JSON(http.StatusOK, views)
-}
-
-// UpsertModelOverride handles PUT /admin/api/v1/model-overrides/{selector}.
-func (h *Handler) UpsertModelOverride(c *echo.Context) error {
-	if h.modelOverrides == nil {
-		return handleError(c, featureUnavailableError("model overrides feature is unavailable"))
-	}
-
-	selector, err := decodeModelOverridePathSelector(c.Param("selector"))
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	var req upsertModelOverrideRequest
-	if err := c.Bind(&req); err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-
-	if err := h.modelOverrides.Upsert(c.Request().Context(), modeloverrides.Override{
-		Selector:  selector,
-		UserPaths: req.UserPaths,
-	}); err != nil {
-		return handleError(c, modelOverrideWriteError(err))
-	}
-
-	override, ok := h.modelOverrides.Get(selector)
-	if !ok || override == nil {
-		slog.Error("model override service returned no override after upsert", "selector", selector)
-		return handleError(c, core.NewProviderError("model_overrides", http.StatusInternalServerError, "model override update failed unexpectedly", nil))
-	}
-	return c.JSON(http.StatusOK, override)
-}
-
-// DeleteModelOverride handles DELETE /admin/api/v1/model-overrides/{selector}.
-func (h *Handler) DeleteModelOverride(c *echo.Context) error {
-	var unavailableErr error
-	var deleteFunc func(context.Context, string) error
-	if h.modelOverrides == nil {
-		unavailableErr = featureUnavailableError("model overrides feature is unavailable")
-	} else {
-		deleteFunc = h.modelOverrides.Delete
-	}
-	return deleteByName(
-		c,
-		unavailableErr,
-		"selector",
-		decodeModelOverridePathSelector,
-		deleteFunc,
-		modeloverrides.ErrNotFound,
-		"model override not found: ",
-		modelOverrideWriteError,
-	)
-}
-
 // ListAuthKeys handles GET /admin/api/v1/auth-keys
-func (h *Handler) ListAuthKeys(c *echo.Context) error {
-	if h.authKeys == nil {
-		return handleError(c, featureUnavailableError("auth keys feature is unavailable"))
-	}
-	views := h.authKeys.ListViews()
-	if views == nil {
-		views = []authkeys.View{}
-	}
-	return c.JSON(http.StatusOK, views)
-}
-
-// CreateAuthKey handles POST /admin/api/v1/auth-keys
-func (h *Handler) CreateAuthKey(c *echo.Context) error {
-	if h.authKeys == nil {
-		return handleError(c, featureUnavailableError("auth keys feature is unavailable"))
-	}
-
-	var req createAuthKeyRequest
-	if err := c.Bind(&req); err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-
-	userPath, err := normalizeUserPathQueryParam("user_path", req.UserPath)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	issued, err := h.authKeys.Create(c.Request().Context(), authkeys.CreateInput{
-		Name:        req.Name,
-		Description: req.Description,
-		UserPath:    userPath,
-		ExpiresAt:   req.ExpiresAt,
-	})
-	if err != nil {
-		return handleError(c, authKeyWriteError(err))
-	}
-	if issued == nil {
-		requestID := strings.TrimSpace(core.GetRequestID(c.Request().Context()))
-		slog.Error("auth key service returned nil issued key", "request_id", requestID, "path", c.Request().URL.Path)
-		return c.JSON(http.StatusInternalServerError, (&core.GatewayError{
-			Type:       core.ErrorType("internal_error"),
-			Message:    "auth key creation failed unexpectedly",
-			StatusCode: http.StatusInternalServerError,
-		}).WithCode("auth_key_issue_failed").ToJSON())
-	}
-	return c.JSON(http.StatusCreated, issued)
-}
-
-// DeactivateAuthKey handles POST /admin/api/v1/auth-keys/:id/deactivate
-func (h *Handler) DeactivateAuthKey(c *echo.Context) error {
-	var unavailableErr error
-	var deactivate func(context.Context, string) error
-	if h.authKeys == nil {
-		unavailableErr = featureUnavailableError("auth keys feature is unavailable")
-	} else {
-		deactivate = h.authKeys.Deactivate
-	}
-	return deactivateByID(c, unavailableErr, "auth key", authkeys.ErrNotFound, "auth key not found: ", deactivate, authKeyWriteError)
-}
-
 // ListAliases handles GET /admin/api/v1/aliases
-func (h *Handler) ListAliases(c *echo.Context) error {
-	if h.aliases == nil {
-		return handleError(c, featureUnavailableError("aliases feature is unavailable"))
-	}
-	views := h.aliases.ListViews()
-	if views == nil {
-		views = []aliases.View{}
-	}
-	return c.JSON(http.StatusOK, views)
-}
-
-// UpsertAlias handles PUT /admin/api/v1/aliases/{name}
-func (h *Handler) UpsertAlias(c *echo.Context) error {
-	if h.aliases == nil {
-		return handleError(c, featureUnavailableError("aliases feature is unavailable"))
-	}
-
-	name, err := decodeAliasPathName(c.Param("name"))
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	var req upsertAliasRequest
-	if err := c.Bind(&req); err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-
-	enabled := true
-	if existing, ok := h.aliases.Get(name); ok && existing != nil {
-		enabled = existing.Enabled
-	}
-	if req.Enabled != nil {
-		enabled = *req.Enabled
-	}
-
-	if err := h.aliases.Upsert(c.Request().Context(), aliases.Alias{
-		Name:           name,
-		TargetModel:    req.TargetModel,
-		TargetProvider: req.TargetProvider,
-		Description:    req.Description,
-		Enabled:        enabled,
-	}); err != nil {
-		return handleError(c, aliasWriteError(err))
-	}
-
-	alias, ok := h.aliases.Get(name)
-	if !ok {
-		return c.NoContent(http.StatusNoContent)
-	}
-	return c.JSON(http.StatusOK, alias)
-}
-
-// DeleteAlias handles DELETE /admin/api/v1/aliases/{name}
-func (h *Handler) DeleteAlias(c *echo.Context) error {
-	var unavailableErr error
-	var deleteFunc func(context.Context, string) error
-	if h.aliases == nil {
-		unavailableErr = featureUnavailableError("aliases feature is unavailable")
-	} else {
-		deleteFunc = h.aliases.Delete
-	}
-	return deleteByName(
-		c,
-		unavailableErr,
-		"name",
-		decodeAliasPathName,
-		deleteFunc,
-		aliases.ErrNotFound,
-		"alias not found: ",
-		aliasWriteError,
-	)
-}
-
 // ListGuardrailTypes handles GET /admin/api/v1/guardrails/types
-func (h *Handler) ListGuardrailTypes(c *echo.Context) error {
-	if h.guardrailDefs == nil {
-		return handleError(c, featureUnavailableError("guardrails feature is unavailable"))
-	}
-	return c.JSON(http.StatusOK, h.guardrailDefs.TypeDefinitions())
-}
-
-// ListGuardrails handles GET /admin/api/v1/guardrails
-func (h *Handler) ListGuardrails(c *echo.Context) error {
-	if h.guardrailDefs == nil {
-		return handleError(c, featureUnavailableError("guardrails feature is unavailable"))
-	}
-	views := h.guardrailDefs.ListViews()
-	if views == nil {
-		views = []guardrails.View{}
-	}
-	return c.JSON(http.StatusOK, views)
-}
-
-// UpsertGuardrail handles PUT /admin/api/v1/guardrails/{name}
-func (h *Handler) UpsertGuardrail(c *echo.Context) error {
-	if h.guardrailDefs == nil {
-		return handleError(c, featureUnavailableError("guardrails feature is unavailable"))
-	}
-
-	name := strings.TrimSpace(c.Param("name"))
-	if name == "" {
-		return handleError(c, core.NewInvalidRequestError("guardrail name is required", nil))
-	}
-
-	var req upsertGuardrailRequest
-	if err := c.Bind(&req); err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-
-	userPath, err := normalizeUserPathQueryParam("user_path", req.UserPath)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	h.mutationMu.Lock()
-	defer h.mutationMu.Unlock()
-
-	if err := h.guardrailDefs.Upsert(c.Request().Context(), guardrails.Definition{
-		Name:        name,
-		Type:        req.Type,
-		Description: req.Description,
-		UserPath:    userPath,
-		Config:      req.Config,
-	}); err != nil {
-		return handleError(c, guardrailWriteError(err))
-	}
-	if err := h.refreshWorkflowsAfterGuardrailChange(c.Request().Context()); err != nil {
-		return handleError(c, err)
-	}
-
-	definition, ok := h.guardrailDefs.Get(name)
-	if !ok {
-		return c.NoContent(http.StatusNoContent)
-	}
-	return c.JSON(http.StatusOK, guardrails.ViewFromDefinition(*definition))
-}
-
-// DeleteGuardrail handles DELETE /admin/api/v1/guardrails/{name}
-func (h *Handler) DeleteGuardrail(c *echo.Context) error {
-	if h.guardrailDefs == nil {
-		return handleError(c, featureUnavailableError("guardrails feature is unavailable"))
-	}
-
-	name := strings.TrimSpace(c.Param("name"))
-	if name == "" {
-		return handleError(c, core.NewInvalidRequestError("guardrail name is required", nil))
-	}
-
-	h.mutationMu.Lock()
-	defer h.mutationMu.Unlock()
-
-	referencingWorkflows, err := h.activeWorkflowGuardrailReferences(c.Request().Context(), name)
-	if err != nil {
-		return handleError(c, err)
-	}
-	if len(referencingWorkflows) > 0 {
-		return handleError(c, core.NewInvalidRequestError("guardrail is used by active workflows: "+strings.Join(referencingWorkflows, ", "), nil))
-	}
-
-	if err := h.guardrailDefs.Delete(c.Request().Context(), name); err != nil {
-		if errors.Is(err, guardrails.ErrNotFound) {
-			return handleError(c, core.NewNotFoundError("guardrail not found: "+name))
-		}
-		return handleError(c, guardrailWriteError(err))
-	}
-	if err := h.refreshWorkflowsAfterGuardrailChange(c.Request().Context()); err != nil {
-		return handleError(c, err)
-	}
-
-	return c.NoContent(http.StatusNoContent)
-}
-
 // ListWorkflows handles GET /admin/api/v1/workflows
-func (h *Handler) ListWorkflows(c *echo.Context) error {
-	if h.workflows == nil {
-		return handleError(c, featureUnavailableError("workflows feature is unavailable"))
-	}
-
-	views, err := h.workflows.ListViews(c.Request().Context())
-	if err != nil {
-		return handleError(c, err)
-	}
-	if views == nil {
-		views = []workflows.View{}
-	}
-	return c.JSON(http.StatusOK, views)
-}
-
-// GetWorkflow handles GET /admin/api/v1/workflows/:id
-func (h *Handler) GetWorkflow(c *echo.Context) error {
-	if h.workflows == nil {
-		return handleError(c, featureUnavailableError("workflows feature is unavailable"))
-	}
-
-	id := strings.TrimSpace(c.Param("id"))
-	if id == "" {
-		return handleError(c, core.NewInvalidRequestError("workflow id is required", nil))
-	}
-
-	view, err := h.workflows.GetView(c.Request().Context(), id)
-	if err != nil {
-		if errors.Is(err, workflows.ErrNotFound) {
-			return handleError(c, core.NewNotFoundError("workflow not found: "+id))
-		}
-		return handleError(c, err)
-	}
-
-	return c.JSON(http.StatusOK, view)
-}
-
-// ListWorkflowGuardrails handles GET /admin/api/v1/workflows/guardrails
-func (h *Handler) ListWorkflowGuardrails(c *echo.Context) error {
-	if h.guardrails == nil {
-		return c.JSON(http.StatusOK, []string{})
-	}
-
-	return c.JSON(http.StatusOK, h.guardrails.Names())
-}
-
-// CreateWorkflow handles POST /admin/api/v1/workflows
-func (h *Handler) CreateWorkflow(c *echo.Context) error {
-	if h.workflows == nil {
-		return handleError(c, featureUnavailableError("workflows feature is unavailable"))
-	}
-
-	var req createWorkflowRequest
-	if err := c.Bind(&req); err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-	scopeProviderName := strings.TrimSpace(req.ScopeProviderName)
-	if scopeProviderName == "" {
-		scopeProviderName = strings.TrimSpace(req.LegacyScopeProvider)
-	}
-	scopeModel := strings.TrimSpace(req.ScopeModel)
-
-	scopeUserPath, err := normalizeUserPathQueryParam("scope_user_path", req.ScopeUserPath)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	scopeProviderName, err = h.validateWorkflowScope(scopeProviderName, scopeModel)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	if err := h.validateWorkflowGuardrails(req.Payload); err != nil {
-		return handleError(c, err)
-	}
-
-	h.mutationMu.Lock()
-	defer h.mutationMu.Unlock()
-
-	version, err := h.workflows.Create(c.Request().Context(), workflows.CreateInput{
-		Scope: workflows.Scope{
-			Provider: scopeProviderName,
-			Model:    scopeModel,
-			UserPath: scopeUserPath,
-		},
-		Activate:    true,
-		Name:        req.Name,
-		Description: req.Description,
-		Payload:     req.Payload,
-	})
-	if err != nil {
-		return handleError(c, workflowWriteError(err))
-	}
-	if version == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	return c.JSON(http.StatusCreated, version)
-}
-
-// DeactivateWorkflow handles POST /admin/api/v1/workflows/:id/deactivate
-func (h *Handler) DeactivateWorkflow(c *echo.Context) error {
-	if h.workflows == nil {
-		return handleError(c, featureUnavailableError("workflows feature is unavailable"))
-	}
-
-	id := strings.TrimSpace(c.Param("id"))
-	if id == "" {
-		return handleError(c, core.NewInvalidRequestError("workflow id is required", nil))
-	}
-
-	h.mutationMu.Lock()
-	defer h.mutationMu.Unlock()
-
-	if err := h.workflows.Deactivate(c.Request().Context(), id); err != nil {
-		if errors.Is(err, workflows.ErrNotFound) {
-			return handleError(c, core.NewNotFoundError("workflow not found: "+id))
-		}
-		return handleError(c, workflowWriteError(err))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-func (h *Handler) refreshWorkflowsAfterGuardrailChange(ctx context.Context) error {
-	if h.workflows == nil {
-		return nil
-	}
-	if err := h.workflows.Refresh(ctx); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (h *Handler) activeWorkflowGuardrailReferences(ctx context.Context, name string) ([]string, error) {
-	if h.workflows == nil {
-		return nil, nil
-	}
-
-	name = strings.TrimSpace(name)
-	if name == "" {
-		return nil, nil
-	}
-
-	views, err := h.workflows.ListViews(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	references := make([]string, 0)
-	for _, view := range views {
-		if !view.Payload.Features.Guardrails {
-			continue
-		}
-		for _, step := range view.Payload.Guardrails {
-			if strings.TrimSpace(step.Ref) != name {
-				continue
-			}
-			references = append(references, view.ScopeDisplay)
-			break
-		}
-	}
-	sort.Strings(references)
-	return references, nil
-}
-
-func (h *Handler) validateWorkflowGuardrails(payload workflows.Payload) error {
-	if !payload.Features.Guardrails || len(payload.Guardrails) == 0 {
-		return nil
-	}
-	if h.guardrails == nil {
-		return featureUnavailableError("guardrail registry is unavailable for workflow authoring")
-	}
-
-	known := make(map[string]struct{}, h.guardrails.Len())
-	for _, name := range h.guardrails.Names() {
-		known[name] = struct{}{}
-	}
-	for _, step := range payload.Guardrails {
-		ref := strings.TrimSpace(step.Ref)
-		if ref == "" {
-			continue
-		}
-		if _, ok := known[ref]; !ok {
-			return core.NewInvalidRequestError("unknown guardrail ref: "+ref, nil)
-		}
-	}
-	return nil
-}
-
-func (h *Handler) validateWorkflowScope(scopeProviderName, scopeModel string) (string, error) {
-	scopeProviderName = strings.TrimSpace(scopeProviderName)
-	scopeModel = strings.TrimSpace(scopeModel)
-
-	if scopeProviderName == "" {
-		if scopeModel != "" {
-			return "", core.NewInvalidRequestError("scope_model requires scope_provider_name", nil)
-		}
-		return "", nil
-	}
-	if h.registry == nil {
-		return "", core.NewInvalidRequestError("provider registry is unavailable for workflow provider-name validation", nil)
-	}
-	if !slices.Contains(h.registry.ProviderNames(), scopeProviderName) {
-		if resolvedProviderName := strings.TrimSpace(h.registry.GetProviderNameForType(scopeProviderName)); resolvedProviderName != "" {
-			scopeProviderName = resolvedProviderName
-		}
-	}
-	if !slices.Contains(h.registry.ProviderNames(), scopeProviderName) {
-		return "", core.NewInvalidRequestError("unknown provider name: "+scopeProviderName, nil)
-	}
-	if scopeModel == "" {
-		return scopeProviderName, nil
-	}
-
-	for _, model := range h.registry.ListModelsWithProvider() {
-		if model.ProviderName == scopeProviderName && model.Model.ID == scopeModel {
-			return scopeProviderName, nil
-		}
-	}
-	return "", core.NewInvalidRequestError("unknown model for provider name "+scopeProviderName+": "+scopeModel, nil)
-}
-
-func decodeAliasPathName(raw string) (string, error) {
-	name, err := url.PathUnescape(strings.TrimSpace(raw))
-	if err != nil {
-		return "", core.NewInvalidRequestError("invalid alias name", err)
-	}
-	name = strings.TrimSpace(name)
-	if name == "" {
-		return "", core.NewInvalidRequestError("alias name is required", nil)
-	}
-	return name, nil
-}
-
-func decodeModelOverridePathSelector(raw string) (string, error) {
-	selector, err := url.PathUnescape(strings.TrimSpace(raw))
-	if err != nil {
-		return "", core.NewInvalidRequestError("invalid model override selector", err)
-	}
-	selector = strings.TrimSpace(selector)
-	if selector == "" {
-		return "", core.NewInvalidRequestError("model override selector is required", nil)
-	}
-	return selector, nil
-}
