+++ internal/admin/routes.go
@@ -1,0 +1,68 @@
+package admin
+
+import "github.com/labstack/echo/v5"
+
+// RouteRegistrar is the subset of *echo.Group / *echo.Echo that RegisterRoutes
+// uses. Decoupling from a concrete echo type keeps the admin package useful for
+// callers that want to mount the API under a different path prefix or wrap the
+// routes with extra middleware.
+type RouteRegistrar interface {
+	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
+	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
+	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
+	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
+}
+
+// RegisterRoutes mounts the admin REST API on the given route group.
+// Callers typically pass an *echo.Group rooted at /admin/api/v1.
+func (h *Handler) RegisterRoutes(g RouteRegistrar) {
+	g.GET("/dashboard/config", h.DashboardConfig)
+	g.GET("/cache/overview", h.CacheOverview)
+
+	g.GET("/usage/summary", h.UsageSummary)
+	g.GET("/usage/daily", h.DailyUsage)
+	g.GET("/usage/models", h.UsageByModel)
+	g.GET("/usage/user-paths", h.UsageByUserPath)
+	g.GET("/usage/log", h.UsageLog)
+	g.POST("/usage/recalculate-pricing", h.RecalculateUsagePricing)
+
+	g.GET("/audit/log", h.AuditLog)
+	g.GET("/audit/conversation", h.AuditConversation)
+
+	g.GET("/providers/status", h.ProviderStatus)
+	g.POST("/runtime/refresh", h.RefreshRuntime)
+
+	g.GET("/budgets", h.ListBudgets)
+	g.PUT("/budgets/:user_path/:period", h.UpsertBudget)
+	g.DELETE("/budgets/:user_path/:period", h.DeleteBudget)
+	g.GET("/budgets/settings", h.BudgetSettings)
+	g.PUT("/budgets/settings", h.UpdateBudgetSettings)
+	g.POST("/budgets/reset-one", h.ResetBudget)
+	g.POST("/budgets/reset", h.ResetBudgets)
+
+	g.GET("/models", h.ListModels)
+	g.GET("/models/categories", h.ListCategories)
+
+	g.GET("/model-overrides", h.ListModelOverrides)
+	g.PUT("/model-overrides/:selector", h.UpsertModelOverride)
+	g.DELETE("/model-overrides/:selector", h.DeleteModelOverride)
+
+	g.GET("/auth-keys", h.ListAuthKeys)
+	g.POST("/auth-keys", h.CreateAuthKey)
+	g.POST("/auth-keys/:id/deactivate", h.DeactivateAuthKey)
+
+	g.GET("/aliases", h.ListAliases)
+	g.PUT("/aliases/:name", h.UpsertAlias)
+	g.DELETE("/aliases/:name", h.DeleteAlias)
+
+	g.GET("/guardrails/types", h.ListGuardrailTypes)
+	g.GET("/guardrails", h.ListGuardrails)
+	g.PUT("/guardrails/:name", h.UpsertGuardrail)
+	g.DELETE("/guardrails/:name", h.DeleteGuardrail)
+
+	g.GET("/workflows", h.ListWorkflows)
+	g.GET("/workflows/guardrails", h.ListWorkflowGuardrails)
+	g.GET("/workflows/:id", h.GetWorkflow)
+	g.POST("/workflows", h.CreateWorkflow)
+	g.POST("/workflows/:id/deactivate", h.DeactivateWorkflow)
+}
