+++ internal/server/http.go
@@ -315,46 +315,7 @@ func New(provider core.RoutableProvider, cfg *Config) *Server {
 
 	// Admin API routes (behind ADMIN_ENDPOINTS_ENABLED flag)
 	if cfg != nil && cfg.AdminEndpointsEnabled && cfg.AdminHandler != nil {
-		adminAPI := e.Group("/admin/api/v1")
-		adminAPI.GET("/dashboard/config", cfg.AdminHandler.DashboardConfig)
-		adminAPI.GET("/cache/overview", cfg.AdminHandler.CacheOverview)
-		adminAPI.GET("/usage/summary", cfg.AdminHandler.UsageSummary)
-		adminAPI.GET("/usage/daily", cfg.AdminHandler.DailyUsage)
-		adminAPI.GET("/usage/models", cfg.AdminHandler.UsageByModel)
-		adminAPI.GET("/usage/user-paths", cfg.AdminHandler.UsageByUserPath)
-		adminAPI.GET("/usage/log", cfg.AdminHandler.UsageLog)
-		adminAPI.POST("/usage/recalculate-pricing", cfg.AdminHandler.RecalculateUsagePricing)
-		adminAPI.GET("/audit/log", cfg.AdminHandler.AuditLog)
-		adminAPI.GET("/audit/conversation", cfg.AdminHandler.AuditConversation)
-		adminAPI.GET("/providers/status", cfg.AdminHandler.ProviderStatus)
-		adminAPI.POST("/runtime/refresh", cfg.AdminHandler.RefreshRuntime)
-		adminAPI.GET("/budgets", cfg.AdminHandler.ListBudgets)
-		adminAPI.PUT("/budgets/:user_path/:period", cfg.AdminHandler.UpsertBudget)
-		adminAPI.DELETE("/budgets/:user_path/:period", cfg.AdminHandler.DeleteBudget)
-		adminAPI.GET("/budgets/settings", cfg.AdminHandler.BudgetSettings)
-		adminAPI.PUT("/budgets/settings", cfg.AdminHandler.UpdateBudgetSettings)
-		adminAPI.POST("/budgets/reset-one", cfg.AdminHandler.ResetBudget)
-		adminAPI.POST("/budgets/reset", cfg.AdminHandler.ResetBudgets)
-		adminAPI.GET("/models", cfg.AdminHandler.ListModels)
-		adminAPI.GET("/models/categories", cfg.AdminHandler.ListCategories)
-		adminAPI.GET("/model-overrides", cfg.AdminHandler.ListModelOverrides)
-		adminAPI.PUT("/model-overrides/:selector", cfg.AdminHandler.UpsertModelOverride)
-		adminAPI.DELETE("/model-overrides/:selector", cfg.AdminHandler.DeleteModelOverride)
-		adminAPI.GET("/auth-keys", cfg.AdminHandler.ListAuthKeys)
-		adminAPI.POST("/auth-keys", cfg.AdminHandler.CreateAuthKey)
-		adminAPI.POST("/auth-keys/:id/deactivate", cfg.AdminHandler.DeactivateAuthKey)
-		adminAPI.GET("/aliases", cfg.AdminHandler.ListAliases)
-		adminAPI.PUT("/aliases/:name", cfg.AdminHandler.UpsertAlias)
-		adminAPI.DELETE("/aliases/:name", cfg.AdminHandler.DeleteAlias)
-		adminAPI.GET("/guardrails/types", cfg.AdminHandler.ListGuardrailTypes)
-		adminAPI.GET("/guardrails", cfg.AdminHandler.ListGuardrails)
-		adminAPI.PUT("/guardrails/:name", cfg.AdminHandler.UpsertGuardrail)
-		adminAPI.DELETE("/guardrails/:name", cfg.AdminHandler.DeleteGuardrail)
-		adminAPI.GET("/workflows", cfg.AdminHandler.ListWorkflows)
-		adminAPI.GET("/workflows/guardrails", cfg.AdminHandler.ListWorkflowGuardrails)
-		adminAPI.GET("/workflows/:id", cfg.AdminHandler.GetWorkflow)
-		adminAPI.POST("/workflows", cfg.AdminHandler.CreateWorkflow)
-		adminAPI.POST("/workflows/:id/deactivate", cfg.AdminHandler.DeactivateWorkflow)
+		cfg.AdminHandler.RegisterRoutes(e.Group("/admin/api/v1"))
 	}
 
 	// Admin dashboard UI routes (behind ADMIN_UI_ENABLED flag)
